\name{RcppURT-package}
\alias{RcppURT-package}
\alias{RcppURT}
\docType{package}
\title{
\packageTitle{RcppURT}
}
\description{
\packageDescription{RcppURT}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RcppURT}
\packageIndices{RcppURT}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{RcppURT}

Maintainer: \packageMaintainer{RcppURT}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
